requirements.txt
nltk~=3.9.1
psycopg2-binary
pyyaml~=6.0.2
textblob~=0.18.0.post0
scikit-learn~=1.0.2  # Cập nhật phiên bản
numpy~=1.21.6  # Cập nhật phiên bản
pandas~=1.3.5  # Cập nhật phiên bản
